requirements.txt
streamlit==1.36.0
langchain==0.2.10
langchain-core==0.2.10
langchain-groq>=0.1.5,<0.2
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
pdfminer.six==20231228
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
httpx==0.27.0
PyYAML==6.0.1
numpy>=1.24.0
